feat: add support for use groups
`pub use a::{b, c, d}` should result in three different
`PublicIdentifier`s being created, but the implementation only
resulted in a single one, so a dummy was being returned. This has
now been fixed to result in a vector of interfaces from a use.

Additionally, the logic is unified to find the public interfaces,
where it was previously duplicated code.

Fixes #4